# Add home-grown `with_na_pass()` for now